Split the functions file into three smaller files, and reduced the number of functions exposed to users
